endpoint/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==6.0.1
PyMySQL==1.1.1
cryptography==44.0.1
pandas==2.2.3
numpy>=1.26.0
yfinance==0.2.28
requests==2.32.4
python-dotenv==1.0.1
APScheduler==3.10.4
ccxt==4.4.24
websocket-client==1.8.0
python-dateutil==2.9.0
pytz==2024.2
PyJWT==2.9.0
Werkzeug==3.0.6
